Deposit Thread complete
Added deposit thread and changed int to double for "amount" in User
class

+++ src/User.java
@@ -41,7 +41,7 @@ public class User {
 
     //**************Methods*************//
     //Function that will deduct "amount" from Balance
-    public void withdraw(int amount)
+    public void withdraw(double amount)
     {
         temp = bankAccount.getBalance();
 
@@ -54,7 +54,7 @@ public class User {
     }
 
     //Function that will add "amount" to Balance
-    public void deposit(int amount)
+    public void deposit(double amount)
     {
         temp = bankAccount.getBalance();
         temp += amount;
